# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AETHERIA E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ──────────────────────────────────────────────────────────────
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get all favorites (by session)
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Favorite"
    post:
      operationId: addFavorite
      tags: [favorites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"
  /favorites/{id}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CART ──────────────────────────────────────────────────────────────────
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/clear:
    delete:
      operationId: clearCart
      tags: [cart]
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cleared
  /cart/apply-coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApply"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"

  # ─── ORDERS ────────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: getCoupons
      tags: [coupons]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROMOTIONS ────────────────────────────────────────────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
...
  /admin/bulk-price-update:
    post:
      operationId: bulkPriceUpdate
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkPriceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpdateResult"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        active:
          type: boolean
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        active:
          type: boolean
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        active:
          type: boolean

    # ── Product ─────────────────────────────────────────────────────────────
    Product:
      type: object
      required: [id, name, price, categoryId]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        isNew:
          type: boolean
        active:
          type: boolean
...
    Favorite:
      type: object
      required: [id, productId, sessionId]
      properties:
        id:
          type: integer
        productId:
          type: integer
        sessionId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
    FavoriteInput:
      type: object
      required: [productId, sessionId]
      properties:
        productId:
          type: integer
        sessionId:
          type: string

    # ── Cart ─────────────────────────────────────────────────────────────────
    Cart:
      type: object
      required: [sessionId, items, subtotal, total]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
    CartItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
        selectedColor:
          type: ["string", "null"]
        price:
          type: number
    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
        selectedColor:
          type: ["string", "null"]
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    CouponApply:
      type: object
      required: [sessionId, code]
      properties:
        sessionId:
          type: string
        code:
          type: string
    CouponValidation:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
        message:
          type: ["string", "null"]

    # ── Order ─────────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minPurchase:
          type: ["number", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        usageLimit:
          type: ["integer", "null"]
        usageCount:
          type: integer
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minPurchase:
          type: ["number", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        usageLimit:
          type: ["integer", "null"]
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minPurchase:
          type: ["number", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        usageLimit:
          type: ["integer", "null"]

    # ── Promotion ────────────────────────────────────────────────────────────
...
    AdminCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [success, token]
      properties:
        success:
          type: boolean
        token:
          type: string
        username:
          type: string
...
    BulkPriceUpdate:
      type: object
      required: [type, value]
      properties:
        type:
          type: string
          enum: [increase_percentage, decrease_percentage, increase_fixed, decrease_fixed]
        value:
          type: number
        categoryId:
          type: ["integer", "null"]
    BulkUpdateResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer
        message:
          type: string