# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AETHERIA E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  shortDescription?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  images?: string[];
  sizes?: string[];
  colors?: string[];
  stock?: number;
  featured?: boolean;
  isNew?: boolean;
  active?: boolean;
}